refactor(server): separate entry script and server handlers
