+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "next build",
-    "start": "node backend/index.js",
-    "dev": "nodemon backend/index.js",
+    "start": "node index.js",
+    "dev": "nodemon index.js",
     "lint": "eslint --fix ."
   },
   "repository": {
